Work Order stability fix - don't fail to send work order if material is missing some fields
Previously I assumed that we knew that the materials would have the expected fields (gender, phenotype etc.).
Since there is a submission bug where they don't, it turns out code building the work order request
(to send to the LIMS) fails when the material is missing some attributes. Now it will send null for missing attributes.