Fixed lock_set_step to set the owner after performing other updates to the new set
The complete/cancel operation is not done in a session for a particular user.
That means that there is no JWT identifying the user to the set service.
That means that if the set is owned by a particular user, this step will error
when it tries to update the set.
Instead, create it without an owner (i.e. owner is "guest"); set its contents;
then lock it and set its owner in the same operation, so no more alterations
can be made.