A155 use billing service to validate and display total price for the work order (#102)
* Added billing facade client
Validation of catalogue using the billing facade to test the product names
Unit tests and integration tests

* Check with the billing service if the cost code is valid when it is selected
while creating a work order

* Only authenticated users can access the product price
Added CSRF header for ajax when getting the product price for the work order
Removed cost_per_sample column from product. This value is going to be obtained from
the billing service depending in a couple of conditions, so we'll store the
value in work order instead.
Added cost_per_sample and total cost to work order, to be able to know the values
presented to the user on work order creation. Both values will be decimal
stored.
BillingFacade client will return a BigDecimal when returning the unit price,
or nil when the unit price is not defined.
Added route to obtain the information of a product inside the scope of a work
order. This is use for adding the information of the cost of the product for
that particular work order. The proposal selected and the product selected
are send to the billing service to obtain the unit cost, and this unit cost
is merged in the product.json obtained.
Added unit tests and integration tests.

* Adds call to tell the billing facade service that the submission is submitted,
completed or cancelled

* Capybara browser in use is not compatible with some new syntax of ES6 and
that was stopping some tests to work.

* Updated from comments