work plan refactors
In work plan status, load all the work orders (so they will be fully cached)
so we get all the information we need in one query instead of several.
Added debugger messages to WorkPlan.project and .original_set, which would
have immediately shown us why everything was so slow when debug logging was enabled.