Broker changes
Broker.working? returns true if everything appears to be ready
for you to send a message. If it is not connected, it tries
to connect. If it IS connected, it checks there are no
unconfirmed messages.
The broker will send an email if it cannot connect or if
there are unconfirmed messages. If will send another if
it tries again later and everything looks ok.
Emails now should go to the user's email address
when you're running in development mode, instead of akerdev.
Emails now include the rails environment.
An instance of event message should now produce the same
json if generate_json is called multiple times.

An instance of broker will now only add one close handler, not one
every time it connects.